refactor: [PersonDetail] SSR 전환
