+++ app/personDetail/components/sideInfo.tsx
