+++ app/personDetail/page.tsx
@@ -1,77 +1,39 @@
-"use client";
-
-import { useContext, useEffect, useState } from "react";
-import { useSearchParams } from "next/navigation";
-
-import { GlobalContext } from "../context";
 import { personDetailApi } from "@/api/httpClient";
 
-import SideInfo from "@/components/pages/personDetail/sideInfo";
-import History from "@/components/pages/personDetail/history";
-import Famous from "@/components/pages/personDetail/famous";
-import Career from "@/components/pages/personDetail/career";
+import SideInfo from "./components/sideInfo";
+import HistoryInfo from "./components/historyInfo";
+import FamousInfo from "./components/famousInfo";
+import CareerInfo from "./components/careerInfo";
 
 import personDetailStyles from "@styles/pages/personDetail/personDetail.module.scss";
 
-export default function PersonDetailPage() {
-  const params = useSearchParams();
-  const { isLoading, setIsLoading } = useContext(GlobalContext);
+async function ServerSideProps(id: any) {
+  const personDetailData = await personDetailApi(id);
 
-  const [biography, setBiography] = useState<string[]>([]);
-  const [famousData, setFamousData] = useState<PersonDetailDataType[]>([]);
-  const [acting, setActing] = useState<PersonDetailDataType[]>([]);
-  const [personInfoData, setPersonInfoData] = useState<personInfoType>({
-    adult: false,
-    also_known_as: [],
-    biography: "",
-    birthday: "",
-    combined_credits: { cast: [], crew: [] },
-    deathday: null,
-    gender: 0,
-    homepage: null,
-    id: 0,
-    imdb_id: "",
-    known_for_department: "",
-    name: "",
-    place_of_birth: "",
-    popularity: 230.142,
-    profile_path: "",
-  });
+  const personInfoData = personDetailData.koData.data;
+  const biography = personDetailData.biography;
+  const famous = personDetailData.famous;
+  const acting = personDetailData.acting;
 
-  useEffect(() => {
-    const personId = params.get("id");
-    setIsLoading(true);
-    personDetailApi(personId).then((res) => {
-      setPersonInfoData(res.koData.data);
-      setBiography(res.biography);
-      setFamousData(res.famous);
-      setActing(res.acting);
-      setIsLoading(false);
-    });
-  }, []);
+  return { personInfoData, biography, famous, acting };
+}
+
+export default async function PersonDetailPage({ searchParams }: any) {
+  const { personInfoData, biography, famous, acting } = await ServerSideProps(
+    searchParams.id
+  );
 
   return (
-    !isLoading && (
-      <div className={personDetailStyles.person_detail}>
-        <SideInfo acting={acting} personInfoData={personInfoData} />
-        <div className={personDetailStyles.detail_info}>
-          <div className={personDetailStyles.person_name}>
-            {personInfoData.name}
-          </div>
-          <div>
-            <div className={personDetailStyles.info_title}>약력</div>
-            <History biography={biography} personInfoData={personInfoData} />
-          </div>
-          <div>
-            <div className={personDetailStyles.info_title}>유명 작품</div>
-            <Famous famousData={famousData} />
-          </div>
-          <div>
-            <div className={personDetailStyles.info_title}>연기</div>
-            <Career acting={acting} />
-          </div>
+    <div className={personDetailStyles.person_detail}>
+      <SideInfo acting={acting} personInfoData={personInfoData} />
+      <div className={personDetailStyles.detail_info}>
+        <div className={personDetailStyles.person_name}>
+          {personInfoData.name}
         </div>
+        <HistoryInfo biography={biography} personInfoData={personInfoData} />
+        <FamousInfo famous={famous} />
+        <CareerInfo acting={acting} />
       </div>
-    )
+    </div>
   );
 }
