+++ styles/pages/personDetail/careerInfo.module.scss
