+++ styles/pages/personDetail/famousInfo.module.scss
