+++ styles/pages/personDetail/historyInfo.module.scss
